Split auth provider verifier tests
